review, add errors, fix logic and move classes
