+++ react/src/components/exposure/gain-panel/NonEditableGainPanel.tsx
@@ -1,4 +1,4 @@
-import { GainType } from "../ExposurePanel.tsx";
+import { GainType } from "./GainPanel.tsx";
 
 type Props = {
   gain: GainType;
