+++ react/src/components/exposure/solve-section/SolvingTabs.tsx
@@ -1,51 +1,17 @@
 import { useState } from "react";
 import { SolveForSNR } from "./SolveForSNR.tsx";
 import { SolveForExposureTime } from "./SolveForExposureTime.tsx";
+import { SimulationSetupParameters } from "../../Simulator.tsx";
+import { ExposureConfigurationType } from "../ExposurePanel.tsx";
 
-export function SolvingTabs({ setupData, update }: any) {
-  const updateWavelength = (newWavelength: any) => {
-    setState({
-      ...state,
-      solveExposureTime: {
-        ...state.solveExposureTime,
-        wavelength: newWavelength,
-      },
-    });
-  };
-  const updateSNR = (newSNR: any) => {
-    setState({
-      ...state,
-      solveExposureTime: {
-        ...state.solveExposureTime,
-        requestedSNR: newSNR,
-      },
-    });
-  };
+type Props = {
+  setup: SimulationSetupParameters;
+  update: (newSetup: ExposureConfigurationType) => void;
+};
 
-  const [state, setState] = useState({
-    solveFor: "signal-to-noise",
-    solveSNR: {
-      exposureTime: 3600,
-      detectorIterations: 1,
-    },
-    solveExposureTime: {
-      requestedSNR: 10,
-      wavelength: 13000,
-    },
-  });
-
+export function SolvingTabs({ setup, update }: Props) {
   const switchToComponent = (componentNumber: number) => {
     setActiveTab(componentNumber);
-    const solveFor =
-      componentNumber === 1
-        ? "Signal To Noise"
-        : componentNumber === 2
-        ? "Exposure Time"
-        : "";
-    update({
-      ...setupData.exposureConfiguration,
-      solveFor,
-    });
   };
 
   const [activeTab, setActiveTab] = useState(1);
@@ -66,16 +32,9 @@ export function SolvingTabs({ setupData, update }: any) {
           </li>
         </ul>
       </div>
-      {activeTab === 1 && <SolveForSNR setupData={setupData} update={update} />}
+      {activeTab === 1 && <SolveForSNR setup={setup} update={update} />}
       {activeTab === 2 && (
-        <SolveForExposureTime
-          setupData={setupData}
-          update={update}
-          solveExposureTime={state.solveExposureTime}
-          updateWavelength={updateWavelength}
-          updateSNR={updateSNR}
-          updatePlot={update}
-        />
+        <SolveForExposureTime setup={setup} update={update} />
       )}
     </>
   );
